+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 - Updates stories for `List` component to use MDX format.
 - Changes `DisplaySizes` to `HeadingDisplaySizes`
+- Enables configurable layout for `Skeleton Loader`
 
 ## 0.23.1